[base] Remove HighFive headers from Storage.h
